consumer-service/requirements.txt
confluent-kafka
google.auth